refactor folder and components title Uppercase
